improve code structure, code cleanup, update Readme
